refactor: moved generateWeeklyTarget function
